💄[Style] 컨트롤러와 뷰 재분리
